qtvcp -fix program hang on different versions of qt
Ok weirdly enough if seems import gtk or pygtk is _not_
nessasary to fix crash.
removing the logging line in offset_origin fixes the problem on version
QT 5.5 and 5.7.
it seems importing pygtk or gtk would allow the logging line error to
pass.
even stranger in different versions of pyqt required import of gtk or import
of pygtk - use the wrong one for the qt version and it errored.
ok... moving on