qtvcp -action buttons can run python code snippets
An action button can be defined to run small snippets of python code
based on state. We might want to limit what it can do..I'll have to look into
that. Using the keyword INSTANCE in in the python_cmd_string allows one
to reference widgets and call handler functions.