qtvcp -add gcode editor widget that actually edits
The gcode_editor can now edit and search through code.
It still needs work to make it friendly and consistant.
It curently display the running gcode - which we might not want..

There is also a gcodeDisplay widget that is just like the old
editor widget - it just displays gcode read only.

We also added a signal to update the percentage of the gcode that has run.
It can be connected to a scrollbar for a visual clue of amount done.